+++ src/seeds/seed-frameworks.ts
@@ -1,21 +1,19 @@
-import { log } from "diginext-utils/dist/xconsole/log";
-
 import type { FrameworkDto, IFramework, IUser, IWorkspace } from "@/entities";
 import { DB } from "@/modules/api/DB";
 
-const initialFrameworks: FrameworkDto[] = [
+export const initialFrameworks: FrameworkDto[] = [
 	{
-		name: "NextJS 13 Starter",
-		repoURL: "https://github.com/digitopvn/next13-starter",
-		repoSSH: "git@github.com:digitopvn/next13-starter.git",
+		name: "Static Site Starter with NGINX",
+		repoURL: "https://github.com/digitopvn/static-nginx-site",
+		repoSSH: "git@github.com:digitopvn/static-nginx-site.git",
 		gitProvider: "github",
 		isPrivate: false,
 		mainBranch: "main",
 	},
 	{
-		name: "Static Site Starter with NGINX",
-		repoURL: "https://github.com/digitopvn/static-nginx-site",
-		repoSSH: "git@github.com:digitopvn/static-nginx-site.git",
+		name: "NextJS 13 Starter",
+		repoURL: "https://github.com/digitopvn/next13-starter",
+		repoSSH: "git@github.com:digitopvn/next13-starter.git",
 		gitProvider: "github",
 		isPrivate: false,
 		mainBranch: "main",
@@ -35,8 +33,6 @@ export const seedFrameworks = async (workspace: IWorkspace, owner: IUser) => {
 			})
 		)
 	).filter((res) => typeof res !== "undefined");
-
-	if (results.length > 0) log(`Workspace "${workspace.name}" > Seeded ${results.length} frameworks.`);
 
 	return results;
 };