+++ src/services/BaseService.ts
@@ -91,6 +91,7 @@ export default class BaseService<T = any> {
 				const userId = user?._id;
 				data.owner = userId;
 
+				if (options.isDebugging) console.log(`${this.model.collection.name} :>> `, user.activeWorkspace);
 				if (this.model.collection.name !== "workspaces" && user.activeWorkspace) {
 					const workspaceId = (user.activeWorkspace as any)._id ? (user.activeWorkspace as any)._id : (user.activeWorkspace as any);
 					data.workspace = workspaceId;